Xcode 6.3 (6D570)

+++ UIKit.framework/iPhoneSimulator.platform/UIKeyboardFlipTransitionView.h
@@ -31,7 +31,7 @@ __attribute__((visibility("hidden")))
 
 @property(nonatomic, getter=isShowingFrontFace) _Bool showingFrontFace; // @synthesize showingFrontFace=_showingFrontFace;
 - (_Bool)transitionIsVisible;
-- (void)rebuildControlKeys:(int)arg1;
+- (void)rebuildControlKeys:(unsigned long long)arg1;
 - (void)rebuildTransitionForSplitStyleChange:(id)arg1;
 - (void)rebuildFromKeyplane:(id)arg1 toKeyplane:(id)arg2 keyboardType:(long long)arg3 orientation:(long long)arg4;
 - (void)updateTransition;
