+++ UIKit.framework/iPhoneSimulator.platform/UIKeyboardSplitTransitionView.h
@@ -42,7 +42,7 @@ __attribute__((visibility("hidden")))
 - (void)transitionToFinalState:(id)arg1;
 - (void)finalizeTransition;
 - (_Bool)transitionIsVisible;
-- (void)rebuildControlKeys:(int)arg1;
+- (void)rebuildControlKeys:(unsigned long long)arg1;
 - (void)rebuildFromKeyplane:(id)arg1 toKeyplane:(id)arg2 keyboardType:(long long)arg3 orientation:(long long)arg4;
 - (void)transformForProgress:(double)arg1;
 - (_Bool)shouldAllowRubberiness;
